+++ plugins/action/custom_issue_definitions_count_info.py
@@ -34,7 +34,7 @@ argument_spec.update(
         name=dict(type="str"),
         priority=dict(type="str"),
         isEnabled=dict(type="bool"),
-        severity=dict(type="float"),
+        severity=dict(type="int"),
         facility=dict(type="str"),
         mnemonic=dict(type="str"),
         headers=dict(type="dict"),
