+++ plugins/action/device_health_info.py
@@ -32,10 +32,10 @@ argument_spec.update(
         deviceRole=dict(type="str"),
         siteId=dict(type="str"),
         health=dict(type="str"),
-        startTime=dict(type="float"),
-        endTime=dict(type="float"),
-        limit=dict(type="float"),
-        offset=dict(type="float"),
+        startTime=dict(type="int"),
+        endTime=dict(type="int"),
+        limit=dict(type="int"),
+        offset=dict(type="int"),
         headers=dict(type="dict"),
     )
 )
