+++ plugins/action/diagnostic_validation_workflows_count_info.py
@@ -29,8 +29,8 @@ argument_spec = dnac_argument_spec()
 # Add arguments specific for this module
 argument_spec.update(
     dict(
-        startTime=dict(type="float"),
-        endTime=dict(type="float"),
+        startTime=dict(type="int"),
+        endTime=dict(type="int"),
         runStatus=dict(type="str"),
         headers=dict(type="dict"),
     )
